requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
qdrant-client==1.9.2
openai==1.45.0
cohere==5.6.2
groq==0.11.0
tiktoken==0.7.0
pydantic==2.9.2
groq>=0.11.0
sentence-transformers